+++ __mocks__/is-online.js
@@ -1,0 +1,3 @@
+const isOnline = jest.fn();
+
+module.exports = isOnline;