+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.31.0] - 2023-08-12
+- s3Uploader introduced to manage branch/repo upload to s3
+- If internet is down, it retries after 5 minutes
+
 ## [3.30.1] - 2023-08-08
 - Fixed passing AWS Config to CloudWatchLogsClient
 - Fixed checking sub directory by checking if parent repo belongs to current user